conhecendo o styled component
