+++ package.json
@@ -25,7 +25,8 @@
     "react-native-reanimated": "^1.10.2",
     "react-native-safe-area-context": "^3.1.4",
     "react-native-screens": "^2.10.1",
-    "react-native-vector-icons": "^7.0.0"
+    "react-native-vector-icons": "^7.0.0",
+    "styled-components": "^5.1.1"
   },
   "devDependencies": {
     "@babel/core": "^7.8.4",
