+++ src/components/Banner/index.js
@@ -1,52 +1,9 @@
 import React, { useState, useEffect, memo } from 'react';
-import {
-  Text,
-  View,
-  StyleSheet,
-  Image,
-  Dimensions,
-  TouchableOpacity,
-  ScrollView,
-} from 'react-native';
+import { Text, View, Image, TouchableOpacity, ScrollView } from 'react-native';
 import Proptypes from 'prop-types';
 import Icon from 'react-native-vector-icons/FontAwesome';
-
-const { width } = Dimensions.get('screen');
-const Style = StyleSheet.create({
-  main: {
-    backgroundColor: 'white',
-    flex: 1,
-    marginRight: 10,
-    marginTop: 20,
-    shadowColor: '#000',
-    shadowOffset: {
-      width: 0,
-      height: 2,
-    },
-    shadowOpacity: 0.25,
-    shadowRadius: 3.84,
-    elevation: 5,
-  },
-  image: { width: width * 0.93, height: 500, alignSelf: 'center' },
-  info: {
-    flex: 1,
-    padding: 5,
-    paddingStart: 10,
-    justifyContent: 'center',
-  },
-  title: { fontSize: 20, fontFamily: 'SourceSansPro-Bold', flex: 2.5 },
-  tagline: {
-    fontSize: 16,
-    marginTop: 10,
-    fontFamily: 'SourceSansPro-Light',
-  },
-  votes: { fontSize: 25, marginTop: 10, fontFamily: 'SourceSansPro-Black' },
-  containertitle: {
-    flexDirection: 'row',
-    marginTop: 10,
-    justifyContent: 'space-between',
-  },
-});
+import Style from './index.style';
+import { Title, Vote } from './styled';
 
 const Banner = ({ movie, hideOverview, onPress, children }) => {
   const [voteAverage, setvoteAvarage] = useState(0);
@@ -87,11 +44,11 @@ const Banner = ({ movie, hideOverview, onPress, children }) => {
         />
         <View style={Style.info}>
           <View style={Style.containertitle}>
-            <Text style={Style.title}>{title}</Text>
+            <Title style={Style.title}>{title}</Title>
             <View style={{ flexDirection: 'row', flex: 1 }}>
               <Text>
                 {' '}
-                <Text style={Style.votes}>{voteAverage} </Text>/ 10
+                <Vote style={Style.votes}>{voteAverage} </Vote>/ 10
               </Text>
               {hasGoodVotes && <Icon name="trophy" size={30} color="#fca311" />}
             </View>
